+++ alloc.h
@@ -10,9 +10,9 @@
 #define MAX_PUSHDATA_REFS (MAX_OPS_PER_SCRIPT * MAX_STACK_SIZE)
 #define ALLOC_PUSHDATA_BYTES (MAX_STACK_SIZE * MAX_OPS_PER_SCRIPT)
 
-struct num * num_pool_new(int *ind);
+struct num * num_pool_new(u16 *ind);
 struct num * num_pool_pop();
-struct num * num_pool_get(const int ind);
+struct num * num_pool_get(const u16 ind);
 
 u8 *byte_pool_new(u32 len, u16 *ind);
 u8 *byte_pool_get(int ind, u32 *len);
